refactor: rename utils to helpers for consistency
